Move fn to util, don't convert id to int when stripping 0
