+++ powergenome/util.py
@@ -930,3 +930,28 @@ def find_region_col(cols: Union[pd.Index, Iterable], context: str = None) -> str
         raise ValueError(s)
     else:
         return region_col[0]
+
+
+def remove_leading_zero(id: Union[str, int]) -> Union[str, int]:
+    """Remove leading zero from IDs that are otherwise integers.
+
+    There is a discrepency between some generator IDs in PUDL and 860m where they are
+    listed with a leading zero in one and an integer in the other. To better match,
+    strip zeros from IDs that would be an integer without them.
+
+    Parameters
+    ----------
+    id : Union[str, int]
+        An integer or string identifier
+
+    Returns
+    -------
+    Union[str, int]
+        Either the original ID (if integer or non-numeric string) or an integer version
+        of the ID with leading zeros removed
+    """
+    if isinstance(id, int):
+        return id
+    elif id.isnumeric():
+        id = id.strip("0")
+    return id
